Create README.md
Data analytics is the technique of analyzing the raw data for making intelligent decisions. For airlines, data analytics is a powerful tool for continuing operations. Airlines Data Analytics for Aviation Industry


Team Member

S.SURYAPRAKASH
P.SANTHOSH 
M.SATHIYAMOORTHI 
S.SIVACHANDHIRAN 

Goal of the Project:

To provide better Airline and AirPort services and to avoid delays in Air Travel across different locations at Municipality level. The aim is to provide airports, airlines, and the travelling public with a neutral, third-party view of which airlines are delivering on their promise to get passengers from Point A to Point B on-time.